WIP: refactor create provider methods to own factory
