+++ src/Core/Client/ClientFactory.php
@@ -97,7 +97,13 @@ class ClientFactory
 
         $options = $this->getDefaultOptions($config);
 
-        $client = $this->createGuzzle6Client($clientClass, $options, $oauthHandler, $logger, $config->getCorrelationIdProvider());
+        $client = $this->createGuzzle6Client(
+            $clientClass,
+            $options,
+            $oauthHandler,
+            $logger,
+            $config->getCorrelationIdProvider()
+        );
 
         if ($client instanceof ContextAwareInterface) {
             $client->setContext($context);
@@ -131,80 +137,6 @@ class ClientFactory
             $cacheAdapterFactory,
             $context
         );
-    }
-
-    public function createTokenStorageProviderFor(Config $config, Client $client, TokenStorage $storage, AnonymousIdProvider $anonymousIdProvider = null)
-    {
-        return $this->createTokenStorageProvider(
-            $config->getOauthUrl(Config::GRANT_TYPE_ANONYMOUS),
-            $config->getOauthUrl(Config::GRANT_TYPE_REFRESH),
-            $config->getClientCredentials(),
-            $client,
-            $storage,
-            $anonymousIdProvider
-        );
-    }
-
-    public function createTokenStorageProvider(
-        $anonTokenUrl,
-        $refreshTokenUrl,
-        ClientCredentials $credentials,
-        Client $client,
-        TokenStorage $storage,
-        AnonymousIdProvider $anonymousIdProvider = null
-    ) {
-        $refreshTokenProvider = $this->createRefreshFlowProvider(
-            $refreshTokenUrl,
-            $credentials,
-            $client,
-            $storage
-        );
-        $anonProvider = $this->createAnonymousFlowProvider(
-            $anonTokenUrl,
-            $credentials,
-            $client,
-            $refreshTokenProvider,
-            $anonymousIdProvider
-        );
-        return new TokenStorageProvider($storage, $anonProvider);
-    }
-
-    public function createPasswordFlowProviderFor(Config $config, Client $client, TokenStorage $storage)
-    {
-        return $this->createPasswordFlowProvider(
-            $config->getOauthUrl(Config::GRANT_TYPE_PASSWORD),
-            $config->getClientCredentials(),
-            $client,
-            $storage
-        );
-    }
-
-    public function createPasswordFlowProvider(
-        $passwordTokenUrl,
-        ClientCredentials $credentials,
-        Client $client,
-        TokenStorage $storage
-    ) {
-        return new PasswordFlowTokenProvider($client, $passwordTokenUrl, $credentials, $storage);
-    }
-
-    public function createAnonymousFlowProvider(
-        $anonTokenUrl,
-        ClientCredentials $credentials,
-        Client $client,
-        RefreshFlowTokenProvider $refreshFlowTokenProvider,
-        AnonymousIdProvider $anonymousIdProvider = null
-    ) {
-        return new AnonymousFlowTokenProvider($client, $anonTokenUrl, $credentials, $refreshFlowTokenProvider, $anonymousIdProvider);
-    }
-
-    public function createRefreshFlowProvider(
-        $refreshTokenUrl,
-        ClientCredentials $credentials,
-        Client $client,
-        TokenStorage $storage
-    ) {
-        return new RefreshFlowTokenProvider($client, $refreshTokenUrl, $credentials, $storage);
     }
 
     private function getDefaultOptions(Config $config)
