+++ tests/integration/ApiTestCase.php
@@ -7,6 +7,7 @@ namespace Commercetools\Core\IntegrationTests;
 use Cache\Adapter\Filesystem\FilesystemCachePool;
 use Commercetools\Core\Client;
 use Commercetools\Core\Client\ClientFactory;
+use Commercetools\Core\Client\ProviderFactory;
 use Commercetools\Core\Client\OAuth\AnonymousIdProvider;
 use Commercetools\Core\Config;
 use Commercetools\Core\Fixtures\InstanceTokenStorage;
@@ -594,7 +595,7 @@ class ApiTestCase extends TestCase
     {
         $config = $this->getClientConfig(['view_products', 'manage_my_profile', 'manage_my_orders', 'manage_my_shopping_lists', 'create_anonymous_token']);
 
-        $provider = ClientFactory::of()->createTokenStorageProviderFor($config, new HttpClient(), new InstanceTokenStorage(), $anonymousIdProvider);
+        $provider = ProviderFactory::of()->createTokenStorageProviderFor($config, new HttpClient(), new InstanceTokenStorage(), $anonymousIdProvider);
         return ClientFactory::of()->createClient($config, $this->getLogger(), $this->getCache(), $provider);
     }
 
@@ -605,10 +606,10 @@ class ApiTestCase extends TestCase
 
         $storage = new InstanceTokenStorage();
         $config = $this->getClientConfig(['view_products', 'manage_my_profile', 'manage_my_orders', 'manage_my_shopping_lists', 'create_anonymous_token']);
-        $provider = ClientFactory::of()->createPasswordFlowProviderFor($config, new HttpClient(), $storage);
+        $provider = ProviderFactory::of()->createPasswordFlowProviderFor($config, new HttpClient(), $storage);
         $provider->getTokenFor($customerDraft->getEmail(), $customerDraft->getPassword());
 
-        $provider = ClientFactory::of()->createTokenStorageProviderFor($config, new HttpClient(), $storage);
+        $provider = ProviderFactory::of()->createTokenStorageProviderFor($config, new HttpClient(), $storage);
         return ClientFactory::of()->createClient($config, $this->getLogger(), $this->getCache(), $provider);
     }
 }
